[refactor]FundingGift 패키지를 domain/gift/entity로 이동 (#${closes})
