+++ src/main/java/org/kakaoshare/backend/domain/funding/repository/FundingGiftRepository.java
@@ -1,6 +1,6 @@
 package org.kakaoshare.backend.domain.funding.repository;
 
-import org.kakaoshare.backend.domain.funding.entity.FundingGift;
+import org.kakaoshare.backend.domain.gift.entity.FundingGift;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface FundingGiftRepository extends JpaRepository<FundingGift, Long> {
