+++ src/main/java/org/kakaoshare/backend/domain/gift/entity/FundingGift.java
@@ -1,18 +1,9 @@
-package org.kakaoshare.backend.domain.funding.entity;
+package org.kakaoshare.backend.domain.gift.entity;
 
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.OneToOne;
+import jakarta.persistence.*;
 import lombok.Builder;
 import lombok.Getter;
-import org.kakaoshare.backend.domain.gift.entity.GiftStatus;
+import org.kakaoshare.backend.domain.funding.entity.Funding;
 
 import java.time.LocalDateTime;
